Move vault edit classes to a separate package
