+++ PayPal/src/main/java/com/braintreepayments/api/paypal/vaultedit/PayPalVaultEditCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.paypal
+package com.braintreepayments.api.paypal.vaultedit
 
 import com.braintreepayments.api.ExperimentalBetaApi
